+++ src/app/test/back-button/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { BackButton } from "@/components/buttons/BackButton";
+import { BackButton } from "@/components/buttons/BackButton/BackButton";
 
 export default function TestBackButton() {
   return <BackButton />;
